pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jaxwavelets"
version = "0.1.0"
description = "JAX-native wavelet transforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Will Handley", email = "wh260@cam.ac.uk"},
]
dependencies = [
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "PyWavelets",
    "ruff",
    "pre-commit",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
    "numpydoc",
]

[project.urls]
Repository = "https://github.com/handley-lab/jaxwavelets"
Documentation = "https://jaxwavelets.readthedocs.io"

[tool.setuptools.packages.find]
include = ["jaxwavelets*"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = [
    "N816",   # mixedCase variables — cA, cD, cH, cV are wavelet convention
    "N802",   # mixedCase function names — same reason
    "N803",   # mixedCase arguments — same reason
    "N806",   # mixedCase in function — same reason
]

[tool.ruff.lint.per-file-ignores]
"jaxwavelets/__init__.py" = ["F401"]      # re-exports are the point of __init__
"benchmarks/*" = ["E402"]            # env vars must be set before jax import

[tool.pytest.ini_options]
testpaths = ["jaxwavelets/tests"]
